requirement.txt
pyyaml
requests
urllib3
pytest
websocket-client
